requirement.txt
phidata
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
openai
python-multipart
wikipedia
sqlalchemy
pgvector
psycopg[binary]